Add background > scripts property for Firefox
Extensions are standardized between Firefox and Chrome except for a few
things such as the way service_workers are definied in the manifest.json.
Firefox wants them in a scripts tag as an array and Chrome in the
"service_worker" tag. Luckily they each ignore the other tag (only
showing a warning). This adds thus adds Firefox support for the extension.